Created a new Contact class and extended the Address,Email and Phone Class from the Contact class
